Hotfix: Temporarily get first byte with GetObject in Registry.headObject (#310)
Tigris Metadata Server is out of sync with the actual state of its
object store. An `aws s3 ls` shows that an object exists but on the
actual download of the object `aws s3 cp`, the object keys sometimes
won't exist.

This is a temporary hotfix to make sure we rebuild the images when an
image does not exist.