Update events logic to allow mapping specific events to specific tags (#333)
HTTP Inputs in Fluent-bit dynamically set `tags` based off the URI of
the request.
e.g. a request to `fluentbit.monitoring:9880/tag_name` results in
fluent-bit automatically setting a tag of `tag_name` to the event

These tag names are then matched with an output in the example below
```
[OUTPUT]
          Name  http
          Match tag_name
          ....
```

By default, the events will map to the tag `http.1` unless it's override
in the input. However, I added a mapping that can be set in
`config.yaml` to map events to specific tags by event name as such:
```
monitoring:
  fluentbit:
    events:
      mapping:
      - name: abstraction.trigger
         tag: internal_api
```