Add callback urls to containers and endpoints (#361)
The way we add callbacks is not ideal but currently there are some
nuances between abstractions that require task callbacks to be built
case by case.
One of these nuances is how we `end_task` in endpoint. It's location of
handling is different than that of task queue, containers, and
functions.
We place `send_callback` after the `end_task` functions.