Multi-Worker node with Multi-GPUs (#394)
Nvidia-smi information relies on file that exists at
`/dev/nvidia[gpuId]`. If you delete this (or if its never mounted), the
user can't see that gpu.
The communication channel of the pod/container to gpu relies on the gpu
bus folder that lives at `/proc/driver/nvidia/gpus/`. Example folder is
`0000:00:06.0` which is the bus id. Without this the user can't
communicate with the gpu. This is mounted at the worker pod level by k8.

Made ContainerCudaManager (now called ContainerNvidiaManager) more
generic through a GPU Manager interface for when we support different
architectures in the future (e.g. AMD).
Pulled out interactions to the GPU's info client out to its own
interface so its easier to mock and test.