Add drop-in config directory (#427)
Since our config library (koanf) allows us to load and overwrite configs
in the order they are loaded, the plan for this is to be able to
generate a config (Secret) via helm and mount it under the drop-in
config directory. The drop-in configs are loaded after the embedded and
CONFIG_PATH configs, but before the CONFIG_JSON config. This makes it so
that the default configs can be overwritten by the drop-in configs, but
the CONFIG_JSON produced by a worker pool still takes precedence.

- Load configs from drop-in directory (/etc/beta9.d/)
- Reuse top-level config in taskqueue and endpoint abstractions
- Don't use global config format when load configs

Resolve BE-1593