Priority worker scheduling (#426)
This PR adds the ability to prioritize workers based on a priority value
set on the worker pool. For example, if priority is `0` (the default) on
both pools, the scheduling logic will select whichever workers are
available in either pool. If one of the pools has a priority of `1`, and
both have available workers, the pool with priority `1` will take
precedence.

**NOTE**: this does not take into account worker creation -- so if there
are _no_ available workers for a particular request, any pool with the
correct GPU will be used to create a new worker. We can handle that case
in a subsequent PR.

- Allow pools to specify a worker priority 
- Clean up resource scheduling/scoring logic