Add deployment download route (#452)
This PR adds a route to the gateway for downloading deployment objects.
It uses the stub ID to generate the download path and returns the file
at that path. This will allow us to build a download option into the UI
so that users can quickly access/verify their deployed code.