Do not prompt user when ending serve (#459)
This change makes it so that serves are terminated immediately on a user
interrupt. The current behavior prompts users to ask if they truly want
to end the serve. However, regardless of the user's input the session
ends. If they chose "n", indicating they wanted the serve to continue,
there will still be a container up that they then need to manually stop.

This change simplifies things and users can easily start a new serve if
they closed the old one on accident.