Add expired status to task and expose task in-queue TTL to sdk (#476)
Task dispatcher already automatically removes task from the list of
pending tasks if it is pending up for 2 hours. However, the `STATUS`
shown is the exact same status as when a `RUNNING` task times out
(`TIMEOUT`). This results in a state that is confusing to the user as
well developers trying to debug issues related to this task activity.