Allow running image build commands in order
Adds new `add_commands()` and `add_python_packages()` to the `Image` abstraction.

This allows for running commands or installing python packages in order, but only after the initial values of the Image's attributes `commands` and `python_packages`. The reason for the exception is to maintain backwards compatibility with existing stubs/deployments.

Resolve BE-1675