Functions in Colab (#566)
This PR fixes functions in Colab. The main issue was that beam was not
found in the last frame in the stack did not include a beam import. As a
result, the beta9 config was being picked up.

To fix this, I added a condition to check if beam is in the sys modules
(its presence there would imply it was imported). I think we might be
able to replace the frame logic entirely with this check. Aside from
that, I've added some error handling to the file syncing and added some
default ignore paths for google drive.