Print error logs in function callable (#587)
Fixes printing error logs in the function callable. It was removed
during a refactor because the error logs were being printed twice, once
in function:main() and once in with_runner_context.

To simplify this, we're decoupling the channel decorators and making it
optional that the upstream error handler prints a stack trace. Also
including a minor change in reordering params for handle_task_failure to
be consistent with complete_task.

Resolve BE-1905