Checkpointing: Cedana runc C/R from worker + CUDA driver level C/R (#624)
## List of changes
- Refactor C/R support to do from the worker instead. A single daemon is
spawned per worker, instead of being spawned inside each runner.
- The runner image has been reverted to initial state. Except that the
base is still Ubuntu 22.04. This is a requirement until we support lower
glibc versions.
- The worker image has been updated to include CRIU, Cedana and
supporting libraries. Uses same CUDA 12.3 image, except that it’s now
Ubuntu 22.04.
- Daemon spawn and health check is done when `NewCedanaClient` is
called. It only succeeds if it was spawned and health check was OK.
- No need of specifying CUDA version through the `—cuda` flag anymore,
as the GPU support is now at the driver API level. As long as the driver
version on the host is supported, C/R should work. Minimum supported
CUDA driver API version is 11.8. Max 12.4 has been tested, but we have
observed it works also with 12.6.
- `DumpStats` and `RestoreStats` are returned as part of
Checkpoint/Restore requests, that include performances statistics.
- The daemon now uses a random free port in the worker, instead of a
fixed one.
- The `base_runc_config.json` has been fully reverted.
- The runc spec is dynamically updated to support C/R (only) in case of
GPU runners. A `/dev/shm` bind mount is added from the worker, as well
as a shared library that is required for GPU interception. This
`/dev/shm` is how the runner process communicates with the GPU
controller (spawned by the daemon) on the worker for GPU C/R.
- Checkpoints are currently stored at `/data/[container_id]`.

## Known issues
- Requires the `/dev/shm` on the worker to be at least 8GiB. We are
already working on making this size dynamic. For testing, use this
example GPU pool config:
```yaml
nvidia: mode: local
      gpuType: any
      runtime: nvidia
      jobSpec:
        nodeSelector: {}
      poolSizing:
        defaultWorkerCpu: 8000m
        defaultWorkerGpuType: ""
        defaultWorkerMemory: 10Gi
        minFreeCpu:
        minFreeGpu:
        minFreeMemory:
        sharedMemoryLimitPct: 100% 
```
- Cedana dynamically removes the `/usr/lib/worker/x86_64-linux-gnu`
mount from spec for now, as it’s causing an issue, since the process is
getting linked with a lib from an external mount. We are investigating
this. Likely, all we will have to do is specify such external mounts in
the `CriuOpts.External` parameter for `JobDump` grpc call.

## TODO improvements
- Currently, when `NewCedanaClient` is called, it blocks until cedana is
fully ready. This can be made asynchronous, as the go routine
`createCheckpoint` anyways waits for the runner to be ready.
- Add an option to toggle compression. We also have a new way of
‘streaming’ checkpoint files directly to a location in `lz4` format,
which is much faster as it does not need multiple disk read/writes.