Feat: Add ability to provide backup GPUs which containers can be scheduled on  (#611)
This update introduces a new feature that allows users to specify
multiple GPUs for fallback purposes. In scenarios where the primary GPU
becomes unavailable due to low capacity or failure, the system will now
automatically fallback to the specified alternative GPUs.

#### Key Changes:
•	GPU Parameter Flexibility:
• The gpu parameter in all abstractions now accepts both a single GPU as
a string and a list of GPUs as strings.
• This change enhances flexibility in GPU management, enabling users to
specify a prioritized list of fallback GPUs.

#### Usage Examples:
1.	Single GPU Specification:
```
@function(
    gpu="A10G",
)
def test():
    ...
```
2.	Multiple GPU Specification:
```
@function(
    gpu=["A10G", "T4"]
)
def test():
    ...
```

#### Priority Handling:
The order of GPUs specified in the list determines their priority. The
GPU at the earliest index in the list is given the highest priority,
allowing for efficient fallback management.

This enhancement aims to improve resource allocation and ensure more
reliable GPU utilization within our system.