Feat: VLLM abstraction (#665)
Resolve BE-1783

This PR adds a VLLM abstraction. Every argument you can provide to `vllm
serve ...` is supported and can be passed into the vllm class or as part
of the `EngineConfig`. The defaults are the same as the ones in vllm's
repo.

Here is a simple example of how it used: 
```python
from beta9 import integrations

e = integrations.EngineConfig()
e.device = "cpu"

vllm_app = integrations.VLLM(name="vllm-abstraction-1", engine_config=e)
```

And we deploy with:
```bash
beta9 deploy app.py:vllm_app
```

Serve with:
```bash 
beta9 serve app.py:vllm_app
```

Below is an example of serving the above on my local machine. 

In order to get deploying/serving a class to work, I introduced a
`set_handler` method. The deploy and serve code now checks if that
message is present and uses it to set the user provided `app:obj`.
Without this, we end up inspecting the sdk to generate the module:obj
and that does not work.

![Screenshot 2024-10-29 at 17 24
47](https://github.com/user-attachments/assets/6fe41b34-8829-4536-a478-9197035f5b05)