Fix: Secrets, autoscaler, and chat templates for vLLM abstraction (#682)
This PR makes two important corrections to the VLLM abstraction. 

First, it adds the secrets parameters. This is critical as it allows us
to pass `HF_TOKEN` as a secret, which is required to download many of
the models.

Second, it adds a chat_template_url parameter. If set, we download this
template and override the chat_template arg.