Fix: Claiming a task and prevent process crashes
When a task cannot be claimed, a none object is returned to the task worker. When this happens, it runs into an Attribute error and causes the process to crash. This prevents the task queue lock and its timeout from being set. This also most likely causes the monitor_task function to keep running, which is probably why we are seeing this error "Task process did not join within the timeout. Terminating..."

Changes

- Use the correct response type when we cannot claim a task
- Fix custom thread pool class exit functionality for Python 3.8
- Fix a few print statements

Resolve BE-1990