Fix: Claiming a task and prevent process crashes (#716)
In TaskQueuePop we claim a task and then update the "task running lock"
TTL. When the task cannot be claimed, we return a nil/None object to the
Python runner when its expecting a TaskQueuePopResponse object. This
causes the spawned Python process to crash. In this scenario,
TaskQueueComplete is never called which is also an issue because the
"keep warm lock" TTL is never updated.

For the AutoscaledInstance of a task queue, the StopContainersFunc has a
pointer to stopContainers which calls stoppableContainers. In
stoppableContainers, a container is considered stoppable when there is
no "keep warm" or "task running" locks. THIS is why a container that
should keep running is ultimately stopped.

A side affect of the Python process crashing is that the monitor_task
function thread keeps running. This is why we see the "Task process did
not join within the timeout. Terminating..." message.

Changes

- Use the correct response type when we cannot claim a task
- Fix custom thread pool class exit functionality for Python 3.8
- Fix a few print statements

Resolve BE-1990