Fix: Send a SIGKILL to container when its really time to stop (#726)
When the goroutine in clearContainer() is unblocked by the select (grace
period has expired or worker receives a SIGTERM), then its really time
to stop the container. In this case, we're going to send a SIGKILL to it
instead of a SIGTERM.

This might fix the issue related to unmounting images.