Fix: Use heartbeats to correct for request token drift (#777)
- Uses key events + request heartbeats to correct for token drift after
a particular gateway crashed unexpectedly

The consequence of this is that if a gateway that was actively handling
a request crashes or is forcibly terminated, there will be a 30 delay
before any active containers that were handling requests have the token
count incremented. This should fix "bricked" containers that had an
inaccurately low token count.