Feat: Allow users to bring their own Dockerfile (#773)
Resolve BE-2084

This feature adds a new operator to the `Image` abstraction that allows
users to specify a path to a Dockerfile that they wish to use as the
base of their Image. We use `buildah` and `umoci` to build their
Dockerfile and gather the bundle needed to run it.

It can be used as follows: 
```python
from beta9 import Image, endpoint

image = Image().from_dockerfile("./Dockerfile").add_python_packages(["numpy"])


@endpoint(image=image, name="test_dockerfile")
def handler():
    return "pass"
```
Where `./Dockerfile` is: 
```
FROM ubuntu:22.04
RUN echo hello
ENV FOO=bar
```

Note: when testing locally, you might need to build gateway again to get
`make start` to work because the go version is bumped.