Fix: fix task cancellation status bug (#794)
In some cases, for a very long running task queue task the dispatcher
can accidentally move a task to "Expired", even though it is actually
completed. This happens because the task is finished, and there is a
race condition in the dispatcher trying to "expire" the task and the
task claim release.