Fix: Remove instance spec from exec env (#802)
For some reason this was causing our base requirement installation to
fail. I suspect it might have been duplicating some of the env vars from
the base spec, which might have been causing issues. We should
investigate adding the vars not already included and see if that still
causes issues. For now, this seems to fix the problem.