Send oom event to es (#826)
Partially resolve BE-2141

This PR makes a small addition to the OOM watching logic. It adds an
additional `PushContainerOOMEvent` to the elastic search event repo. It
is called when OOMs occur. In the frontend, we can use the container ID
and stub ID to alert users that one of their containers has OOM issues.