Copy custom images to disk instead of memory before unpacking (#831)
Resolve BE-2204

This PR adds additional logging to custom base image builds, increases
the spin up timeout for build containers, and copies images to disk
instead of memory with skopeo. The biggest change is the copying
location. This change has little to no impact on the time it takes to
copy/unpack custom images from what I've measured. It is a change worth
making because it decreases the memory required by the worker for custom
base images. For example, as things are,
`docker.io/pytorch/pytorch:2.2.2-cuda12.1-cudnn8-devel` will cause the
worker to OOM. With this change, that image can be used without any
problems.