Fix: Stuck build containers (#836)
Resolve BE-2203

This PR contains to work to fix one of the likely causes of stuck build
workers. Basically, if a user cancels a build (Ctrl-C) before the build
container has started, the worker will get stuck. As things are, the
entire build will carry on as if nothing happened and then the container
will just sit there forever.

To prevent this from happening we can use redis pub/sub. When the
context is cancelled in the gateway, we publish a stop-build event to a
channel containing the container id (`stop-build-{container_id}`).
Meanwhile, when the worker claims the container request, it subscribes
to that same channel. If an event comes through it cancels the context
that has been passed down.