Feat: Use cloudpickle to run remote functions in jupyter environments (#856)
Resolve BE-2218

This PR changes the strategy used for running remote functions in
notebooks. Instead of creating a temporary file built by stitching
together executed cells, we use cloudpickle to capture the required
function + deps. This resolves issues with re-executed cells and is
generally more reliable.

This PR also changes the way that the default python environment is
selected for `Image`. It tries to see if the active python version from
the local environment is a valid remote python version. If it is, it
will use that version by default. I can remove this change or limit it
to ipython environments only. This is nice when working in jupyter
notebooks because cloudpickle dumps are dependent on the python version.
If you create a dump in 3.12 and try to load it in 3.10, things will not
work.


TODO: 
- [x] Test whether different protocols improve compatibility for
cloudpickle python version mixing
- [x] Alert users when they try to run different local/remote python
versions in jupyter environments
- [ ] Explore ways to detect function content changes to avoid
re-syncing when the cell has been re-executed but no code has changed