Feat: Set build container spin up timeout based on image size (#884)
Resolve BE-2217

I pulled the skopeo logic out of the image client and into its own
struct. This makes it easy to use the inspect functionality in the
gateway's builder. Aside from that, I do a simple calculation of 120
nanoseconds per byte. That evaluates to 2 minutes per GB of the base
image. That is a very generous amount of time.