Fix: don't overwrite /root in containers (#890)
Certain images install stuff into /root. We currently overwrite this
path with a tmpfs on container startup. I don't think this is needed,
because we are already adding a r/w overlay on top of the user container
that allows them to modify the contents. For example, this should work:

```
from beta9 import Image, endpoint


@endpoint(
    cpu=1,
    memory=128,
    image=Image(
        python_version="python3.10",
    ).add_commands(
        [
            "touch /root/hello.txt",
            "touch /something.txt",
        ]
    ),
)
def handler():
    with open("/root/hello.txt", "r") as f:
        print(f.read())

    with open("/root/something_else.txt", "w") as f:
        f.write("Hello, World!")

    return "Hello, World!"
```