Feat: repository services (Pt. 1) (#892)
Adds the go2proto package into our protobuf generation script, and
automatically generates the protobuf message types. Any struct with the
// @go2proto comment above them in types will be exported as protobuf
messages.

This is part 1, which lays out the various structs and sets up the grpc
services for the worker repo and container repo. The follow up will
implement the actual repository services needed for the worker with the
generated types.