Feat: Detect existing python (#912)
Resolve BE-2249

If a user does not specify a python version, we should first see if the
image they are using contains one before installing 3.10. Many images do
(pytorch, jax, miniconda, etc) and we currently just install a new
version and mess them up.

Note: this only works once this fix is in
https://github.com/beam-cloud/beta9/pull/905