fix: add RetrieveLock support to redislock (#917)
Due to the way our distributed lock works (redislock), when it obtains a
lock, it associates that lock with a particular host using a token. In
other words, if I acquire a lock somewhere, it's expected that I will
release it after I'm done in the that same host.

The problem is -- if you put that locking mechanism behind an API, you
don't know which host is going to need to _release_ that lock. What this
PR does is allow us to return that lock token (which was already part of
the internal mechanics of redislock), so that a different host than who
created the lock gets the `Release` call, it can recreate the lock using
the provided token (if the lock is still held), and then release it.