fix: juicefs does not exit when the worker is oomkilled (#936)
When resource limits/requests are configured, the worker gets a
oom_score_adj of -997 (or not -1000), but all juicefs processes get
-1000 which means they are not oom-killable.

This is mentioned in the
[oom_score_adj](https://man7.org/linux/man-pages/man5/proc_pid_oom_score_adj.5.html)
man page.

> The value of oom_score_adj is added to the badness score
> before it is used to determine which task to kill.
> Acceptable values range from -1000 (OOM_SCORE_ADJ_MIN) to
> +1000 (OOM_SCORE_ADJ_MAX).  This allows user space to
> control the preference for OOM-killing, ranging from always
> preferring a certain task or completely disabling it from
> OOM-killing.  The lowest possible value, -1000, is
> equivalent to disabling OOM-killing entirely for that task,
> since it will always report a badness score of 0.

Changes to fix this:

- Match parent's oom_score_adj for all juicefs processes. This value is
dynamically adjusted based on a pod's QoS which is based on resources
limits/requests. This ensures juicefs gets the same oom_score_adj as the
worker.
- Don't use daemon mode for juicefs because it immediately creates a
defunct process since it's started in a goroutine anyway

Resolve BE-2334