Feat: add deployment id to container list if available (#977)
Requires an extra query for each container, but for CLI usage it should
be fine. Better would be to cache this ID in state somewhere to avoid
the extra queries. Querying by stub ID should be safe to look up the
deployment since we `force_create_stub` when preparing the runtime for
any new deployment -- so it should be one-to-one

---------

Co-authored-by: Luke Lombardi <luke@beam.cloud>