Fix: Nil Autoscaler on older stub configs (#981)
Move the forceful override of a nil autoscaler to the `Sync` method. 
The sync method is called on initial initialization of the autoscaled
instance (NewAutoscaledInstance) as well as Load method in the
InstanceController.

We must have the nil autoscaler override snippet every time we do a sync
because it will reload the StubConfig from the database anew.