fix: run juicefs in foreground within goroutine
This re-adds part of the previous juicefs process improvements. This should not run as a daemon because the daemon ends up forking another juicefs process and becomes defunct immediately. When the worker is shutting down, it can't shutdown the forked juicefs process because the daemonized/parent juicefs process is already shutdown/zombied.