Feat: Storage refactor (#1020)
Modifies storage architecture to move away from a single JFS mount and
towards a WorkspaceStorage model. There are a few pieces still missing
from this, but I've opened it up for review.

**TODO**
- [x] Migrate output stat / creation to work in both the workspace
storage model and legacy storage model
- [x] Modify volume services to allow listing via StorageClient instead
of direct mounts (conditionally, if workspace storage is enabled)
- [ ] Fix symlinks in geesefs
- [x] Clean up / ensure stub object upload works for legacy workspaces
as well as new workspaces. This may require a couple of small changes
- [x] Implement geeseFS unmount logic
- [x] Implement fuse mount cleanup by workspace (opportunistically,
after a grace period) in worker. It is currently done naively on worker
shutdown
- [x] Review config changes and flatten anything that can be flattened

---

These need to be done in a follow up:
- [ ] Automatically create buckets for users on workspace creation
- [ ] Prepare migration script to create workspace storage objects and
migrate data

---------

Co-authored-by: Luke Lombardi <luke@beam.cloud>