Add pool and host name in logs for adding and removing requests from workers (#1080)
Currently we just log that we are adding `request_id` to `worker_id`.
This information is not particularly useful as we have no idea where
`worker_id` actually is. With this change, we add the pool and hostname,
so that we can easily figure out exactly where the request is actually
going from the log.