Fix: Dont try to load empty result (#1124)
If a function times out, the sdk blows up on the user because it tries
to load an empty result.

Long term, we should add an exit code and render a warning in the
terminal, but this will do for now.
    
<!-- This is an auto-generated description by mrge. -->
---

## Summary by mrge
Fixed a bug where the SDK would crash if it tried to load an empty
result after a function timeout.

<!-- End of auto-generated description by mrge. -->