tweak k3d yaml (#1151)
This small change just makes local development on a linux machine
easier. There are permissions issues when building the worker once
`.k3d` in the project directory. They re-appear everytime a PVC is
created or modified, so easiest to just avoid having it in that
directory.
    
<!-- This is an auto-generated description by mrge. -->
---

## Summary by mrge
Updated k3d.yaml to use a directory in the home folder instead of the
project directory for storage and registry volumes. This avoids
permission issues during local development on Linux.

<!-- End of auto-generated description by mrge. -->