Feat: connect shell to any existing container (#1223)
This PR adds the ability to connect `beta9 shell` to any existing
container using
`beta9 shell --container-id [container-id]`

I replaced openssh-server with dropbear which is a lightweight SSH
server that is compiled into a single binary that listens on port 2222.
On a regular container, this server is not spun up by default and
requires `beta9 shell` to initialize it.

Once the dropbear server is spun up, it persists until the container
spins down.