Dont start container on delayed worker when build is cancelled  (#1250)
resolve BE-2680


There is an edge case where the build worker has not yet started when
the build is cancelled. In these cases the event is lost and the worker
starts the container.

This change sets the status of the build container to "STOPPING" when
its cancelled. We can then check that status and cancel the build
context on the worker side.